+++ ReleaseNotes.md
@@ -181,6 +181,8 @@ Please change your configuration according to [App.xaml](https://github.com/spie
 * Consider `ResizeMode` for the window caption buttons
 * `ResizeGrip` in `MaterialWindow`
 * New control `TransitionContentControl`
-* Optional custom content for dialogs
+* New dialog features
+  * Optional custom content for dialogs
+  * Control for simple input dialogs
 ### Fixes
 * Fixed reflection code in `ResourceDictionaryExtensions`